Update README.md
This basic calculator, built with HTML, CSS, and JavaScript, offers a user-friendly interface for performing fundamental arithmetic operations (addition, subtraction, multiplication, and division). Users can input numbers and operators by clicking on the on-screen buttons, with the entered expression displayed dynamically. The calculator features a clear button, an equals button to display results, and additional functionalities for changing the sign of the current input and calculating percentages. The project includes cal.html for structure, style.css for styling, and script.js for the interactive functionality.